restructure components folders
