+++ components/users/index.ts
@@ -1,0 +1,2 @@
+export * from './users-table-placeholder';
+export * from './users-table';
