+++ components/users/users-table-placeholder.tsx
@@ -1,6 +1,6 @@
-import RefreshButton from './refresh-button';
+import { RefreshButton } from '../buttons';
 
-export default function UsersTablePlaceholder() {
+export function UsersTablePlaceholder() {
   return (
     <div className="bg-white/30 p-12 shadow-xl ring-1 ring-gray-900/5 rounded-lg backdrop-blur-lg max-w-xl mx-auto w-full">
       <div className="flex justify-between items-center mb-4">
